pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Abjust", email = "2533898108@qq.com"},
]
dependencies = [
    "graia-ariadne[standard]>=0.11.7",
    "ntplib>=0.4.0",
    "numpy>=1.25.2",
    "requests>=2.31.0",
]
requires-python = ">=3.11,<4.0"
readme = "README.md"
license = {text = "AGPL-3.0"}

